Initial test YAML file
Tests are stored under the t/ directory. 1.yaml is intended to be a
valid (i.e. boolean true, i.e. 1) YAML file. For the time being, it is
simply a hand-crafted collection of both standard and "pathological"
YAML syntax. This just gives us something to munch on during
development.

At the moment, we are not adopting a TDD approach, as this project's
primary motivation pedagogical. However, if the parser gets off the
ground, then we will likely want to consider more robust testing
strategies.