a.apln: Recommended cleanup and conventions
Selecting multiple indices with Squad, unfortunately, requires enclosing
them. Roger Hui has lamented this [citation needed] and demonstrated the
utility of an `I` function that mimics J's From (`{`) verb.

This also adopts Roger Hui's `assert`, and Aaron Hsu's convention of
`IN` for the input array in a linear dataflow architecture.