a.apln: Remove "pretty-print" function
Dyalog treats CR as the "normal" line break, i.e. a carriage return,
followed by a line feed. LF, on the other hand is treated simply as a
line feed, meaning that LF-separated lines will trail off endlessly to
the right when printed in a Dyalog session.

The `PP` function was just a nicety to fixup the LF characters with CR
for printing. This kind of general development helper belongs in a
dyalog startup Load[0] script more than inside the parser directly.

[0]:http://help.dyalog.com/18.2/#UserGuide/Installation%20and%20Configuration/Configuration%20Parameters/Load.htm