code: Break long line across mulitple lines
One of the purported benefits of APL, especially the style we employ
here, is that it puts much more on the screen at one time, enabling a
direct view of macro features in the architecture. As Arthur Whitney
once famously quipped, "I hate scrolling."

Long lines that run out of the viewport directly run against such a
philosophy. Here, we stick with the 80 character width standard limit.
Not only is this a well-established default, it meshes nicely with
research on readability vs paragraph-width[0].

[0]:http://web.archive.org/web/20170918212943/http://psychology.wichita.edu/surl/usabilitynews/72/LineLength.asp