code: Move notes etc. out of parser body
In this project we are experimenting with a style that emphasizes global
perspective and macroscopic grokkability over traditional "readability"
evaluations.

Previously, some notes and "comment code" was interspersed within the
body of P, but this goes against the above organizational objectives. In
fact, the fact that these notes spread out the code cause a mild feeling
of annoyance for me. That's an interesting data point in the experiment.