P: Tweak spelling of depth-normalization idiom
Conceptually, there are two changes to the scan function:

1. Commute arguments

Conceptually, the scan is a left-associative one, so by commuting ⍺
contains the left-head and ⍵ contains the next right-element.

2. Pre-normalize arguments to rank 1

In the most general case, both ⍺ and ⍵ are rank 1 arguments to Interval
Index. For the scan, ⍵ is always rank 0 but can be lifted to rank 1
without semantic change. However, in the first application of scan, ⍺ is
rank 0 which is incompatible with Interval Index. Thus, by pre-raveling
the arguments with Over, we normalize their ranks to the general case.