P: Normalize escaped quotes in double-quoted strings
In order to detect double-quoted strings, we needed to start
tracking node kinds. Doing the same for implicit node tags
is also similar, so this commit adds that as well.

In addition, so that changes to the Y namespaces get picked up by
tests, we had to change the way names are reflected into T. Previously,
we were copying the definitions from Y into T, which meant that
changes to the definitions in Y wouldn't get picked up in T until
the UTILS namespace was reloaded. Since Link has no way of knowing
this, changes to PS, for example, wouldn't get reflected in tests.

Instead, we just create a pointer to the Y namespace from within T.