Rewrite from a clean slate
After developing the test framework in the previous commits, it became clear
that the complexity/code ratio of PS was too high for me to confidently
navigate around and make changes to the code. It felt undirected and
disorganized.

This commit is an attempted rebirth, attempting to keep the following salient:

1. TDD: Incrementally externalize domain knowledge in tests
2. Delay semantic reification as late as possible